Move type definitions to utils #4
