+++ src/utils.ts
@@ -6,3 +6,26 @@ export function isValidUrl(urlString: string): boolean {
     return false;
   }
 }
+export type SolarIrradianceData = {
+  metadata: {
+    description: string;
+    latitude: number;
+    longitude: number;
+    samples_phi: number;
+    samples_theta: number;
+  };
+  data: Array<{
+    theta: number;
+    phi: number;
+    radiance: number;
+  }>;
+};
+export type SphericalPoint = {
+  altitude: number;
+  azimuth: number;
+};
+export type Point = {
+  x: number;
+  y: number;
+  z: number;
+};
